Add benchmark for remove operation (#135)
Part of #121 - Added benchmark for remove operation.
For benchmarks a separate Maven project in subdirectory of this project was created. I'm not sure about including ppom in parent for pom.xml in benchmarks but without it qulice is not included.
Readme will be updated in next PR.
Also removed todo from Helm class as it was done in previous PRs.